+++ app/src/main/java/ru/s4m1d/notes/app/desktop/client/app/runner/AppStarter.java
@@ -4,9 +4,9 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import ru.s4m1d.notes.app.desktop.client.controller.context.SwingApplicationContext;
-import ru.s4m1d.notes.app.desktop.client.controller.context.NotesContext;
-import ru.s4m1d.notes.app.desktop.client.controller.context.WorkspaceContext;
+import ru.s4m1d.notes.app.desktop.client.model.context.SwingApplicationContext;
+import ru.s4m1d.notes.app.desktop.client.model.context.NotesContext;
+import ru.s4m1d.notes.app.desktop.client.model.context.WorkspaceContext;
 import ru.s4m1d.notes.app.desktop.client.controller.task.MainFrameStateService;
 import ru.s4m1d.notes.app.desktop.client.controller.task.context.ContextInitializationTask;
 import ru.s4m1d.notes.app.desktop.client.controller.notes.NotesMonitoringServiceImpl;
