--- controller-api/src/main/java/ru/s4m1d/notes/app/desktop/client/controller/datamodel/content/Content.java
@@ -1,5 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.controller.datamodel.content;
-
-public interface Content<T> {
-    public T get();
-}
