--- controller-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/controller/task/workspace/WorkspaceService.java
@@ -1,26 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.controller.task.workspace;
-
-import lombok.AllArgsConstructor;
-import ru.s4m1d.notes.app.desktop.client.controller.context.WorkspaceContext;
-import ru.s4m1d.notes.app.desktop.client.components.workspace.WorkSpaceLayeredPane;
-
-import javax.swing.text.DefaultStyledDocument;
-import java.io.IOException;
-import java.io.InputStream;
-
-@AllArgsConstructor
-public class WorkspaceService {
-    private WorkSpaceLayeredPane workSpaceLayeredPane;
-    private WorkspaceContext workspaceContext;
-
-    public void loadContent() {
-        try {
-            workSpaceLayeredPane.getWorkspaceScrollPane().getTextEditorPane().read(
-                    (InputStream) workspaceContext.getCurrentNote().getContent().get(),
-                    new DefaultStyledDocument()
-            );
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
