+++ daemon-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/controller/task/MainFrameStateService.java
