+++ model/src/main/java/ru/s4m1d/notes/app/desktop/client/model/content/Content.java
@@ -1,0 +1,5 @@
+package ru.s4m1d.notes.app.desktop.client.model.content;
+
+public interface Content<T> {
+    public T get();
+}
