+++ model/src/main/java/ru/s4m1d/notes/app/desktop/client/model/context/SwingApplicationContext.java
@@ -1,0 +1,13 @@
+package ru.s4m1d.notes.app.desktop.client.model.context;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class SwingApplicationContext {
+    private WorkspaceContext workspaceContext;
+    private NotesContext notesContext;
+}
